Emojis removed and bot stays in vc when idle now
Removed emojis and removed the "bot leaves instantly when queue ends" feature. Just using this bot for my personal discord and my friends specifically asked for the bot to stay when queue leaves so when people are playing one song at a time we don't hear the join/leave notifications everytime.